+++ src/data-sources/likes.js
@@ -1,10 +1,9 @@
 const { DataSource } = require('apollo-datasource');
-const Sequelize = require('sequelize');
 
 class LikesAPI extends DataSource {
-  constructor() {
+  constructor({ store }) {
     super();
-    this.store = createStore();
+    this.store = store;
   }
 
   async getMovieLikes({ user }) {
@@ -27,33 +26,5 @@ class LikesAPI extends DataSource {
     return !!like;
   }
 }
-
-const createStore = () => {
-  const sequelize = new Sequelize('database', 'username', 'password', {
-    dialect: 'sqlite',
-    logging: false,
-    // storage: './.store.sqlite',
-    operatorsAliases: false,
-  });
-
-  const CREATE_LIKES_QUERY = `CREATE TABLE likes(
-    id INTEGER PRIMARY KEY,
-    createdAt DATETIME,
-    updatedAt DATETIME,
-    user TEXT,
-    movie TEXT
-  )`;
-
-  sequelize.query(CREATE_LIKES_QUERY);
-
-  const likes = sequelize.define('like', {
-    user: Sequelize.STRING,
-    movie: Sequelize.STRING,
-  });
-  
-  likes.sync({ alter: true, force: true });
-
-  return { likes };
-};
 
 module.exports = LikesAPI;
